requirements.txt
# Image pipeline (Module 3)
python-dotenv>=1.0.0
roboflow>=1.1.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
pandas>=2.0.0

# Audio pipeline (Module 1)
openai-whisper
spacy>=3.7.0
transformers>=4.36.0
torch>=2.0.0

# Text pipeline (Module 5)
nltk>=3.8.0

# Document pipeline (PDF / Word)
python-docx>=1.1.0
pdfplumber>=0.11.0
pymupdf>=1.24.0
